requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.0
tldextract>=5.1.0
tqdm>=4.66.0
